requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database & Cache
psycopg2-binary>=2.9.9
django-redis==5.4.0
django-filter==25.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography>=41.0.0
bcrypt==4.0.1

# API Documentation
drf-spectacular==0.27.0

# Content Management
django-ckeditor-5==0.2.12
django-taggit==5.0.1
Pillow==10.4.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Validation & Serialization
marshmallow==3.20.1
jsonschema==4.19.2

# External APIs
requests==2.31.0
openai==1.3.6
beautifulsoup4==4.12.2
# lxml==4.9.3  # Commented out due to compilation issues
html5lib==1.1  # Alternative HTML parser

# Monitoring & Logging
sentry-sdk==1.38.0
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0